+++ src/app/components/spinner/index.tsx
@@ -1,5 +1,8 @@
 import React from "react";
-import { ActivityIndicator, StyleSheet, View } from "react-native";
+import { ActivityIndicator, View } from "react-native";
+import { TURQUOISE } from '@constants/colors'
+
+import styles from './styles';
 
 const withSpinner = (Component: any) => (props: any) => {
   const {isLoading, ...otherProps} = props;
@@ -7,24 +10,12 @@ const withSpinner = (Component: any) => (props: any) => {
     <>
     {isLoading ?
       <View style={[styles.container, styles.horizontal]}>
-        <ActivityIndicator size="large" color="#0000ff" />
+        <ActivityIndicator size="large" color={TURQUOISE} />
       </View> :
       <Component {...otherProps} />
     }
     </>
   )
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center"
-  },
-  horizontal: {
-    flexDirection: "row",
-    justifyContent: "space-around",
-    padding: 10
-  }
-});
 
 export default withSpinner;
