+++ src/app/components/spinner/styles.ts
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center"
+  },
+  horizontal: {
+    flexDirection: "row",
+    justifyContent: "space-around",
+    padding: 10
+  }
+});
+
+export default styles;
